switch to integer currency (#2)
no longer saving expenses as floats. instead a Currency class is
used to act as a proxy between what is shown and what is saved in
the database. both parts of the amount are signed. not completely
intuitive but integer 0 is not signed => have to save the sign in
cents if necessary.